[F] Add numcolors option
It allows to pick a given number of colors from the colormap, regardless
of the number of columns used. Additionally, one can use 'D[0..n]' as a
color to select one of this colors extracted from the colormap.

Add test and baseline images for tests covering Y scale on the right.